my_electron_django/requirements.txt
numpy
pyyaml
pytz
pyinstaller==5.12.0
django
